+++ hslua-packaging/test/HsLua/Packaging/DocumentationTests.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE TypeApplications #-}
 {-|
 Module      : HsLua.Packaging.DocumentationTests
-Copyright   : © 2021-2024 Albert Krewinkel
+Copyright   : © 2021-2026 Albert Krewinkel
 License     : MIT
 Maintainer  : Albert Krewinkel <tarleb@hslua.org>
 
@@ -11,10 +11,10 @@ Tests for calling exposed Haskell functions.
 module HsLua.Packaging.DocumentationTests (tests) where
 
 import Data.Version (makeVersion)
-import HsLua.Core (top, Status (OK), Type (TypeNil, TypeString))
+import HsLua.Core (top, Status (OK))
 import HsLua.Packaging.Documentation
 import HsLua.Packaging.Function
-import HsLua.Marshalling (forcePeek, peekIntegral, pushIntegral, peekText)
+import HsLua.Marshalling (peekIntegral, pushIntegral)
 import Test.Tasty.HsLua ((=:), shouldBeResultOf)
 import Test.Tasty (TestTree, testGroup)
 
@@ -23,16 +23,14 @@ import qualified HsLua.Core as Lua
 -- | Calling Haskell functions from Lua.
 tests :: TestTree
 tests = testGroup "Documentation"
-  [ testGroup "Function docs"
-    [ "retrieves function docs" =:
-      "factorial" `shouldBeResultOf` do
+  [ testGroup "getdocumentation"
+    [ "retrieves function docs as userdata" =:
+      Lua.TypeUserdata `shouldBeResultOf` do
         pushDocumentedFunction factorial
-        _ <- getdocumentation top
-        TypeString <- Lua.getfield top "name"
-        forcePeek $ peekText top
+        getdocumentation top
 
     , "returns nil for undocumented function" =:
-      TypeNil `shouldBeResultOf` do
+      Lua.TypeNil `shouldBeResultOf` do
         OK <- Lua.dostring "return function () return 1 end"
         getdocumentation top
     ]